# gate the cast before movements in lazy